+++ docs/src/App.vue
@@ -1,15 +1,34 @@
-<template>  
-  <Layout />
+<script>
+//import HelloWorld from './components/HelloWorld.vue'
+
+//export default {
+//  name: 'App',
+//  components: {
+//    HelloWorld
+//  }
+//}
+</script>
+
+<template>
+  <el-config-provider namespace="ep">
+    <BaseHeader />
+    <div style="display: flex">
+      <BaseSide />
+      <div>
+        <img alt="Vue logo" class="element-plus-logo" src="./assets/img/logo.png" />
+        <HelloWorld msg="Hello Vue 3.0 + Element Plus + Vite" />
+      </div>
+    </div>
+  </el-config-provider>
 </template>
 
-<script>
 
-import Layout from './layouts/Layout.vue'
-
-export default {
-  name: 'App',
-  components: {
-    Layout
-  }
+<style>
+#app {
+  text-align: center;
+  color: var(--ep-text-color-primary);
 }
-</script>
+.element-plus-logo {
+  width: 50%;
+}
+</style>